# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Order & Promotion API",
        "description": "REST API for browsing products, managing a cart, applying promotional codes, checking out and managing orders.\n\n**Money** is returned as decimal strings (e.g. `\"149.99\"`) and accepted in dollars; it is stored as integer cents.\n\n**Errors** share one shape: `{ \"message\", \"code\", \"details\"? }`. Validation errors (`validation_failed`) add an `errors` object keyed by field.\n\n**Authentication**: call `POST /api/auth/login` (or register), then send `Authorization: Bearer <token>`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "This server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List my orders",
                "description": "The customer's own orders, newest first.",
                "operationId": "listOrders",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/Page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Your orders, newest first.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show one of my orders",
                "description": "Orders belonging to other customers return 403 (`forbidden`).",
                "operationId": "showOrder",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order id.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The order with its items.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List active products",
                "description": "Filtering and sorting use the spatie/laravel-query-builder syntax. Unknown filters or sorts return 422. Results are cached per filter combination and refreshed whenever a product or its stock changes.",
                "operationId": "listProducts",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Case-insensitive partial name match; %, _ and commas are matched literally.",
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    },
                    {
                        "name": "filter[min_price]",
                        "in": "query",
                        "description": "Minimum price in dollars (inclusive).",
                        "schema": {
                            "type": "string",
                            "example": "10.00"
                        }
                    },
                    {
                        "name": "filter[max_price]",
                        "in": "query",
                        "description": "Maximum price in dollars (inclusive); must be at least min_price.",
                        "schema": {
                            "type": "string",
                            "example": "50.00"
                        }
                    },
                    {
                        "name": "filter[in_stock]",
                        "in": "query",
                        "description": "true: stock above 0; false: out of stock.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false",
                                "1",
                                "0"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Prefix with - for descending. Default -created_at.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "price",
                                "-price",
                                "name",
                                "-name",
                                "created_at",
                                "-created_at"
                            ]
                        }
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/Page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated active products.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show an active product",
                "description": "Inactive products return 404 like missing ones.",
                "operationId": "showProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "required": [
                    "message",
                    "code"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Only 3 units of 'Desk Lamp' are available."
                    },
                    "code": {
                        "description": "Machine-readable error code.",
                        "type": "string",
                        "example": "insufficient_stock"
                    },
                    "details": {
                        "description": "Extra context for some errors.",
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "required": [
                    "message",
                    "code",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The quantity field must be at least 1."
                    },
                    "code": {
                        "type": "string",
                        "example": "validation_failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "quantity": [
                                "The quantity field must be at least 1."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 4
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Walnut Desk Lamp"
                    },
                    "sku": {
                        "type": "string",
                        "example": "LAMP-001"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "string",
                        "example": "49.99"
                    },
                    "stock_quantity": {
                        "type": "integer",
                        "example": 5
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "OrderItem": {
                "description": "Snapshot taken at purchase time; later product changes do not affect it.",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "Null if the product was later deleted.",
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Walnut Desk Lamp"
                    },
                    "sku": {
                        "type": "string",
                        "example": "LAMP-001"
                    },
                    "unit_price": {
                        "type": "string",
                        "example": "49.99"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "line_total": {
                        "type": "string",
                        "example": "99.98"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "subtotal": {
                        "type": "string",
                        "example": "154.00"
                    },
                    "discount": {
                        "type": "string",
                        "example": "30.80"
                    },
                    "total": {
                        "type": "string",
                        "example": "123.20"
                    },
                    "promotion_code": {
                        "type": "string",
                        "example": "SUMMER20",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "cancelled_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthenticated": {
                "description": "Missing or invalid bearer token (`unauthenticated`).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Unauthenticated.",
                            "code": "unauthenticated"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Authenticated but not allowed (`forbidden`), e.g. another customer's order or a non-admin on an admin route.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "You do not have access to this order.",
                            "code": "forbidden"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "The resource does not exist (`product_not_found`, `order_not_found`, `cart_item_not_found`, `promotion_not_found`, `not_found`).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Product not found.",
                            "code": "product_not_found"
                        }
                    }
                }
            },
            "ValidationFailed": {
                "description": "The input is invalid (`validation_failed`).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "Rate limit exceeded (`too_many_requests`); see the `Retry-After` header.",
                "headers": {
                    "Retry-After": {
                        "description": "Seconds until another request is allowed.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Too Many Attempts.",
                            "code": "too_many_requests"
                        }
                    }
                }
            }
        },
        "parameters": {
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Items per page (1–100).",
                "schema": {
                    "type": "integer",
                    "default": 15,
                    "maximum": 100,
                    "minimum": 1
                }
            },
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Page number.",
                "schema": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token from /api/auth/login or /api/auth/register.",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}